Refactor WooPosSearchByIdentifier to improve product search handling with dedicated classes for GTIN and SKU searches
