+++ WooCommerce/src/main/kotlin/com/woocommerce/android/ui/woopos/common/data/searchbyidentifier/WooPosSearchByIdentifierException.kt
@@ -1,0 +1,3 @@
+package com.woocommerce.android.ui.woopos.common.data.searchbyidentifier
+
+internal class WooPosSearchByIdentifierException(val error: WooPosSearchByIdentifierResult.Error) : Exception()
